fix(003-admin-center): 统一错误码到 spec + 修复登录跳回登录页
T102：把后端业务错误码全部对齐到 spec.md §Error Codes：
  1001 wrong password / 1002 account disabled / 1003 account locked /
  1004 token invalid|expired / 2001 insufficient permission /
  3001 admin not found / 3002 phone already exists /
  3003 cannot delete super admin / 3004 cannot disable last super admin
  4000/4040/4090/5000 作为无 spec 码时的兜底
- crates/hiveweb/src/utils/error.rs：以 `codes` 常量 + 语义化
  AppError 变体替换原来的 HTTP-status 形式（40101/40301/...）
- crates/hiveweb/src/api/auth.rs / admin.rs：每个 handler 走对应
  的语义化 AppError
- crates/hiveweb/src/middleware/auth.rs：缺 / 过期 / 错误 JWT 改返回
  ApiResponse(code=1004, http=401)，而非裸 401

会话调试中顺手修复的两处实战 bug：
- services/auth.rs::is_account_locked 把 Redis TTL 收成 usize。
  Redis 对不存在 key 返回 -2，-2 as usize 在 64 位下变 usize::MAX-1，
  恒大于 0 → 任何登录都被判 "locked"。改用 i64，并注释 -2/-1/>0 三种语义
- web/src/services/api.ts：在 axios 响应拦截器里 unwrap 后端
  {code, message, data} envelope，避免所有 service 都吃到整个 envelope；
  此前导致 useAuth.login 拿到 response.token === undefined，
  setToken("undefined") → 紧接着 /auth/me 401 → 拦截器把 token 清掉
  → "登录成功后立即跳回登录页"
- web/src/components/__tests__/AdminTable.test.tsx：补 updated_at 让
  fixture 通过 tsc 类型检查

+++ crates/hiveweb/src/services/auth.rs
@@ -59,9 +59,14 @@ pub async fn get_failed_attempts(redis: &RedisClient, phone: &str) -> Result<i64
 }
 
 pub async fn is_account_locked(redis: &RedisClient, phone: &str) -> Result<bool> {
+    // Redis TTL semantics:
+    //   -2 → key does not exist
+    //   -1 → key exists but has no expiry (counter is still being accumulated below the lockout threshold)
+    //   >0 → seconds remaining until expiry (account is locked)
+    // Must use a signed type so the negative sentinels survive the wire decode.
     let key = format!("{}{}", LOGIN_FAILED_KEY_PREFIX, phone);
     let mut conn = redis.get_multiplexed_async_connection().await?;
-    let ttl: usize = conn.ttl(&key).await?;
+    let ttl: i64 = conn.ttl(&key).await?;
     Ok(ttl > 0)
 }
